change2

+++ public/javascripts/product.js
@@ -2,14 +2,20 @@ $(function () {
     Array.from($('.del')).forEach(i =>
         i.onclick = function () {
             let id = $(i).parent().parent().attr("datakey");
-            $.ajax({
-                url: "/table1/del/" + id,
-                type: "DELETE",
-                success: function () {
-                    alert("删除成功");
-                    $(i).parent().parent().remove();
-                }
-            });
+            let r = confirm('是否确认删除?');
+            if (r == true) {
+                $.ajax({
+                    url: "/table1/del/" + id,
+                    type: "DELETE",
+                    success: function () {
+                        alert("删除成功");
+                        $(i).parent().parent().remove();
+                    }
+                });
+            } else {
+                alert('取消删除');
+            }
+
         });
 });
